// src/views/media/Device/Playback/typings.d.ts
export type recordsItemType = {
    channelId: string;
    deviceId: string;
    endTime: number;
    fileSize: number;
    name: string;
    secrecy: string;
    startTime: number;
    mediaEndTime: number;
    mediaStartTime: number;
    filePath: string;
    type: string;
    id: string;
    isServer?: boolean;
};